Added obocreator script
This script takes a set of files (the parsed data from reddit, or users
in the future), and inserts new terms for each word into the ontology.

Currently, there are no properties for terms (such as
healthy/unhealthy), and only one term can be associated with a concept,
which is conceptually great as it simplifies the knowledge domain.