Added diet terms (from reddit data) to ontology
Note: I removed several terms from each ontology that I considered
unnecessary as there were complex dishes (often containing various
concepts), or not suited for this ontology, i.e. beurre, which is
French for butter, or pie, which without context it’s difficult to
determine the concept — it could be a meat or vegetable pie.