Added messenger & initial OEQs
Messenger selects an appropriate OEQ response to send to the user based
on the message received. This is based on a word-frequency of the
user’s message to detect the most frequent concepts in the user message
in comparison with the ontology (stored in SQL db).

Currently, the OEQ is selected at random from the list obtained for the
concept, for example, if weight-lifting is the most common concept
detected in the users message then there are five possible responses.
One of those is selected at random to send to the user.

Understandable, this is not a good way to send questions to a user (as
they may not be context appropriate). An improved and alternative
solution is being developed (based on user feedback), which will happen
after the interface and initial system is developed.