Use buttons (arrows) for feedback.
- Added glyph icons to use arrows, though I feel this is quite bloated
given I am just using two of the arrows. Alas, it is provided by
bootstrap, which also feels quite bloated for this application… I may
remedy this later!

- NOTE: I plan to use the arrows to enable the user to ‘switch’ the
message, and provide feedback for the OEQ sent, i.e. when they press an
arrow the stored value will incremented/de-incremented depending on the
arrow direction (left is decrease, right is increase).