Use stemming to improve concept selection
Using snowballStemmer as it’s an improvement over Porter, and much less
aggressive than Lancaster.

Stemming improves the selection process as it can detect words in the
ontology that are similar to those stemmed, e.g. ‘appl’ from ‘apples’
where only ‘apple’ exists in the ontology.

Of course, plurals and synonyms could be saved into the ontology, but
this approach removes such complexity.