Address bias in open-ended question selection
Previously, the system would only select the top-rated questions, so,
for example, if there was 5 questions, and 2 were rated higher (say .9,
others .6), then only the 2 questions would be used!

To address this, I have opted to use all the highest rated questions,
then other questions that have not been sent (regardless of rating —
this is to improve the system and diversify questions sent). Finally,
once all questions have been sent, then a question is selected from all
at random.

To achieve this, I have added an additional field (conversation_id) to
the messages table that will store the unique id of each session. This
can then be used to determine what messages have been sent.

The messages sent/received are not saved in the database.