Simplify how data is stored/read
- Store the contents of each category in a single file rather than
multiple, and set the filename to be date rather than folder. This
simplifies the code, and reading by users.
- Further, I have removed the frequency from files as it’s not used
anywhere for analysis, and would be simple to re-add should it be
required in the future.
- A new file (known_words) exists that stores words parsed by
‘obocreator’ that do not meet our ontology requirements. This allows us
to not be asked again to categorise these words, which is the manual
labor in ontological development… sigh!